Add context manager API to HandlerSubscriptionAPI
