+++ newsfragments/989.feature.rst
@@ -1,0 +1,1 @@
+The ``HandlerSubscriptionAPI`` now supports a context manager interface, removing/cancelling the subscription when the context exits
